Add draft connection functionality.
We are connected to the database in this iteration. Clients are able to join the room associated with the draft. There are checks for both the correct lobby code and side code in the server.ts websocket. code is a bit long so these will change locations in the future. I had to comment out nearly all stat backend code because it was outdated and gave errors. This will be fixed in the stats v1.5 update