Fix repeat bans and picks
I have no clue how this was only happening clientside. I suspect it had to do with event listeners not being properly handled with the clients constantly reloading. Will need to revisit this later but for now the temporary fix just ends the function early if the champion was already banned/picked.