Refactor how state is handled serverside
This all stemmed from wanting to have the custom names displayed. This allows for the actual use of the database only for tournament code drafts. Others will only be saved as a record and live on the server. The draft record is now initialized at the creation of the lobby, instead of when someone uses the link to join.