Refactor state management
This makes the server create and update a state variable specifically for the client. The clintside also has a draft state state variable, which takes the place of the 20 state variables I had before. The UI is seperated from the draft function as well (somewhat) on the client, so goodbye 400+ line file