Refactor draft tool
Massive backend update in preperation for adding new/optional draft phases. The sockets within the draftSocket function have had their functions split into their own files (draftHandlers) and all of the phase handlers have been put into that folder. The nested if statements that handled the phase changes have been refactored to its own function as a switch statement. the phases are in an array that can be added to in order to support new phases.